delete files that are no longer needed, clean code from comments, add function that gets tensors from csv
