+++ all_data.csv
@@ -29,3 +29,132 @@ time,price
 1706775390420,42142.4609375
 1706775410601,42152.64453125
 1706775430771,42153.69140625
+1706786254559,42252.26171875
+1706786274920,42251.76171875
+1706786295065,42251.76171875
+1706786315240,42253.390625
+1706786333928,42253.390625
+1706786356062,42248.7578125
+1706786376618,42249.02734375
+1706786396752,42247.1484375
+1706786416912,42247.1484375
+1706786437032,42247.53125
+1706786457133,42247.53125
+1706786477246,42243.4296875
+1706786712147,42217.578125
+1706786732318,42216.41796875
+1706786752456,42210.79296875
+1706786772780,42210.79296875
+1706786793037,42208.74609375
+1706786813135,42216.65234375
+1706787324752,42233.265625
+1706787344963,42233.265625
+1706787365177,42230.50390625
+1706787385323,42233.046875
+1706787405498,42233.046875
+1706787425682,42240.87109375
+1706787445870,42246.98828125
+1706787466397,42246.98828125
+1706787486508,42251.3359375
+1706787506667,42257.0078125
+1706787526850,42253.1875
+1706787547012,42253.1875
+1706787567203,42249.9765625
+1706787587307,42256.11328125
+1706787607510,42256.11328125
+1706787627679,42257.53515625
+1706787647886,42256.43359375
+1706787668000,42256.43359375
+1706787688208,42253.7578125
+1706787708386,42253.7578125
+1706787728877,42251.98828125
+1706787749052,42252.5390625
+1706787769225,42252.5390625
+1706787789390,42244.98046875
+1706787809604,42245.86328125
+1706787830130,42245.86328125
+1706787850327,42245.640625
+1706787870512,42236.4296875
+1706787890664,42236.4296875
+1706787910845,42237.5703125
+1706787931789,42232.58984375
+1706787951983,42232.58984375
+1706787972096,42233.23046875
+1706787992247,42233.23046875
+1706788012532,42231.578125
+1706788032933,42218.57421875
+1706788053061,42218.57421875
+1706788073249,42201.6015625
+1706788093939,42190.1796875
+1706788114104,42190.1796875
+1706788134277,42181.88671875
+1706788154445,42183.73046875
+1706788174630,42183.73046875
+1706788194797,42180.2890625
+1706788215063,42180.2890625
+1706788235295,42177.265625
+1706788255400,42172.2109375
+1706788275505,42172.2109375
+1706788295685,42159.83984375
+1706788315778,42151.59375
+1706788335900,42151.59375
+1706788355999,42141.58203125
+1706788376171,42128.55078125
+1706788396297,42128.55078125
+1706788416548,42120.171875
+1706788436664,42120.171875
+1706788456764,42118.453125
+1706788476923,42125.89453125
+1706788497077,42125.89453125
+1706788517194,42127.46875
+1706788537332,42118.1171875
+1706788557511,42118.1171875
+1706788577998,42121.04296875
+1706788598167,42121.04296875
+1706788618323,42120.84375
+1706788638435,42130.08203125
+1706788658569,42130.08203125
+1706788678875,42125.87109375
+1706788699021,42124.515625
+1706788719315,42117.6484375
+1706788739482,42117.6484375
+1706788759653,42091.9765625
+1706788779816,42087.7109375
+1706788799918,42087.7109375
+1706788820172,42085.08203125
+1706788840339,42085.08203125
+1706788860716,42079.1171875
+1706788880950,42082.5078125
+1706790078958,42127.0703125
+1706790099245,42127.0703125
+1706790119351,42128.48828125
+1706790139432,42128.48828125
+1706790159518,42125.85546875
+1706790179669,42125.63671875
+1706790199759,42125.63671875
+1706790219856,42125.4140625
+1706790239952,42124.47265625
+1706790260059,42124.47265625
+1706790280155,42115.6171875
+1706790300268,42110.484375
+1706790320361,42110.484375
+1706790340512,42109.2734375
+1706790360601,42118.296875
+1706790380688,42118.296875
+1706790400792,42121.80078125
+1706790420883,42121.80078125
+1706790440962,42126.984375
+1706790461087,42127.39453125
+1706790481194,42125.44140625
+1706790501281,42125.44140625
+1706790521371,42132.1953125
+1706790541463,42142.4375
+1706790561557,42142.4375
+1706790581728,42145.74609375
+1706790601839,42154.32421875
+1706790622005,42154.32421875
+1706790642086,42158.625
+1706790662177,42161.20703125
+1706790682285,42161.20703125
+1706790702379,42159.38671875
+1706790722466,42159.38671875
