--- data.csv
@@ -1,5 +1,0 @@
-time,price
-1.706718992053e12,43430.38771764296
-1.706718999449e12,43430.38771764296
-1.706719053329e12,43415.04301897931
-1.706719135428e12,43410.45153764589
