+++ lib/crypto_price_fetcher.ex
@@ -16,8 +16,9 @@ defmodule CryptoPriceFetcher do
   defp parse_price(body) do
     with {:ok, data} <- Jason.decode(body),
          %{"data" => %{"priceUsd" => price_usd}, "timestamp" => timestamp} <- data do
-      current_time_utc = DateTime.utc_now()
-      local_time = DateTime.shift_zone(current_time_utc, "Europe/Paris", Tzdata.TimeZoneDatabase)
+      # current_time_utc = DateTime.utc_now()
+
+      # local_time = DateTime.shift_zone(current_time_utc, "Europe/Paris", Tzdata.TimeZoneDatabase)
       {:ok, price_usd, timestamp}
     else
       error -> {:error, error}
@@ -29,14 +30,9 @@ defmodule CryptoPriceFetcher do
       IO.puts("ID")
       IO.inspect(self())
       current_price_float = String.to_float(current_price)
-      df_all = Explorer.DataFrame.from_csv!("all_data.csv")
-      map = Explorer.DataFrame.to_series(df_all)
-      old_x_vals = Explorer.Series.to_list(map["time"])
-      old_y_vals = Explorer.Series.to_list(map["price"])
-      updated_tensor_x = concatenate_number_to_tensor(Nx.tensor([old_x_vals]), time)
-
-      updated_tensor_y =
-        concatenate_number_to_tensor(Nx.tensor([old_y_vals]), current_price_float)
+      {x, y} = get_tensors_from_csv()
+      updated_tensor_x = concatenate_number_to_tensor(x, time)
+      updated_tensor_y = concatenate_number_to_tensor(y, current_price_float)
 
       print_tensors(updated_tensor_x, updated_tensor_y)
       save_tensors_to_files(updated_tensor_x, updated_tensor_y, "all_data.csv")
@@ -69,7 +65,7 @@ defmodule CryptoPriceFetcher do
   end
 
   defp concatenate_number_to_tensor(tensor, number) do
-    updated_tensor = Nx.concatenate([tensor, Nx.tensor([[number]])], axis: 1)
+    updated_tensor = Nx.concatenate([tensor, Nx.tensor([[number]])])
     updated_tensor
   end
 
@@ -78,5 +74,27 @@ defmodule CryptoPriceFetcher do
     a = Explorer.Series.from_tensor(tensor_x)
     df = Explorer.DataFrame.new(%{time: a, price: b})
     Explorer.DataFrame.to_csv(df, file)
+  end
+
+  def get_tensors_from_csv() do
+    df_all = Explorer.DataFrame.from_csv!("all_data.csv")
+    map = Explorer.DataFrame.to_series(df_all)
+    old_x_vals = Explorer.Series.to_list(map["time"])
+    old_y_vals = Explorer.Series.to_list(map["price"])
+    IO.puts("old")
+    IO.inspect(old_x_vals)
+    old_tensors_x = list_to_tensors(old_x_vals)
+    IO.puts("new")
+    IO.inspect(old_tensors_x)
+    old_tensors_y = list_to_tensors(old_y_vals)
+    tensor_x = Nx.tensor(old_tensors_x)
+    IO.puts("tensor")
+    IO.inspect(tensor_x)
+    tensor_y = Nx.tensor(old_tensors_y)
+    {tensor_x, tensor_y}
+  end
+
+  defp list_to_tensors(list) do
+    Enum.map(list, fn x -> [x] end)
   end
 end
